requirements.txt
asgiref==3.6.0
Django==3.0.14
django-rest-framework==0.1.0
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==18.9.0
hangul-utils==0.2
mysqlclient==2.1.1
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
